Errata for fix bug #565470 entitled "Compression/decompression failure with large IPv6 extension".
Add the files with the sizes of the generated ROHC packets for the new
non-regression test.

+++ src/comp/c_generic.h
@@ -305,7 +305,7 @@ struct list_comp
 	int (*get_index_table)(const struct ip_packet *ip, const int index);
 
 	/// @brief the handler used to get the size of an extension
-	int (*get_size)(const unsigned char *ext);
+	unsigned short (*get_size)(const unsigned char *ext);
 
 	/// @brief the handler used to compare two extension of the same type
 	int (*compare)(const unsigned char *ext,
